Initial removal of dmon.
Moved around more of dmon, and made a lua class that'll handle most of the stuff.

Getting to a bootable point.

At a good spot.

Changed load order for plugins, made it so that we know about ignore as soon as possible.

Removed print statement.

Changed windows functioanlity over to something that is usable.

Removed threads dependency.

Fixed init.